requirements.txt
Django==5.2.6
djangorestframework>=3.13,<4
django-cors-headers==4.0.0
djangorestframework-simplejwt>=5.0,<6  # JWT authentication
psycopg2-binary # For PostgreSQL
gunicorn # Production WSGI server
whitenoise # For serving static files

# Optional for GIS support
# psycopg2-binary is already listed; ensure GDAL/GEOS installed system-wide
